# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Moderation Bot Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface AuthUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatar?: string | null;
  avatarUrl: string;
}

export interface Guild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  memberCount: number;
  botPresent: boolean;
}

export interface GuildDetail {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  memberCount: number;
  /** @nullable */
  description?: string | null;
}

export interface Channel {
  id: string;
  name: string;
  type: number;
}

export interface LogChannelMap {
  /** @nullable */
  member_ban?: string | null;
  /** @nullable */
  member_unban?: string | null;
  /** @nullable */
  member_timeout_add?: string | null;
  /** @nullable */
  member_timeout_remove?: string | null;
  /** @nullable */
  member_kick?: string | null;
  /** @nullable */
  member_join?: string | null;
  /** @nullable */
  member_leave?: string | null;
  /** @nullable */
  member_nickname_change?: string | null;
  /** @nullable */
  member_role_add?: string | null;
  /** @nullable */
  member_role_remove?: string | null;
  /** @nullable */
  member_voice_move?: string | null;
  /** @nullable */
  member_voice_disconnect?: string | null;
  /** @nullable */
  message_delete?: string | null;
  /** @nullable */
  message_edit?: string | null;
  /** @nullable */
  role_create?: string | null;
  /** @nullable */
  role_delete?: string | null;
  /** @nullable */
  role_update?: string | null;
  /** @nullable */
  channel_update?: string | null;
  /** @nullable */
  channel_permissions_update?: string | null;
  /** @nullable */
  invite_create?: string | null;
  /** @nullable */
  command_used?: string | null;
  /** @nullable */
  server_update?: string | null;
  /** @nullable */
  member_warn?: string | null;
}

export interface GuildConfig {
  guildId: string;
  /** @nullable */
  defaultLogChannel?: string | null;
  logChannels: LogChannelMap;
  /** @nullable */
  creationVoiceChannel?: string | null;
  updatedAt?: string;
}

export interface GuildConfigInput {
  /** @nullable */
  defaultLogChannel?: string | null;
  logChannels?: LogChannelMap;
  /** @nullable */
  creationVoiceChannel?: string | null;
}

export interface GuildStats {
  guildId: string;
  totalMembers: number;
  membersJoinedLast24h: number;
  membersLeftLast24h: number;
  messagesLast24h: number;
  /** @nullable */
  onlineMembers?: number | null;
}

export interface MemberHistoryPoint {
  date: string;
  joins: number;
  leaves: number;
  total: number;
}

export interface MessageHistoryPoint {
  date: string;
  count: number;
}

export type LogEntryMetadata = { [key: string]: unknown };

export interface LogEntry {
  id: number;
  guildId: string;
  eventType: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  targetId?: string | null;
  description: string;
  metadata?: LogEntryMetadata;
  createdAt: string;
}

export interface LevelingConfig {
  guildId: string;
  textXpMin: number;
  textXpMax: number;
  textXpCooldown: number;
  voiceXpPerMinute: number;
  /** @nullable */
  levelupChannelId?: string | null;
  levelupMessageEnabled: boolean;
}

export interface LevelingConfigInput {
  textXpMin?: number;
  textXpMax?: number;
  textXpCooldown?: number;
  voiceXpPerMinute?: number;
  /** @nullable */
  levelupChannelId?: string | null;
  levelupMessageEnabled?: boolean;
}

export interface LeaderboardEntry {
  userId: string;
  xp: number;
  level: number;
  rank: number;
}

export interface BotInvite {
  url: string;
}

export type ListGuildChannelsParams = {
/**
 * Discord channel type filter: 0=text, 2=voice, 5=announcement. Omit to get text+announcement.
 */
type?: number;
};

export type ListGuildModerationParams = {
/**
 * Filter actions targeting a specific user ID
 */
userId?: string;
};

export type GetLeaderboardParams = {
/**
 * Which leaderboard: text (default) or voice
 */
category?: GetLeaderboardCategory;
};

export type GetLeaderboardCategory = typeof GetLeaderboardCategory[keyof typeof GetLeaderboardCategory];


export const GetLeaderboardCategory = {
  text: 'text',
  voice: 'voice',
} as const;